requirements.txt

pyserial~=3.5